refactor: switch to bincode for encoding requests
Replace serde_json with bincode for serializing and deserializing
requests. Comes with the downside of reduced introspectablity in the
case of broken data ending up in the database but massively reduces the
size of an encoded message.

BREAKING CHANGE: The Error struct has had its serialization error
variant changed from Json containing a serde_json error to Serde
containing a bincode error