feat: Add accepting ranges of responses
Change the type in the set of accepted response codes for a request to
an enum which representents either a category of response codes, a range
of response codes, or a single response code. This makes it easier to
accept a large set or an entire category of response codes, which
previously would have been very tedious.

BREAKING CHANGE: change the type of the accept_responses field in the
Request struct