feat: Add spawning jobs with custom settings
Add job spawning methods which accept a closure that can be used to
change settings for a job, such a whether it should be ordered and the
amount of retry attempts. Also change jobs to being ordered by default,
since it's less problematic to forget disabling ordered jobs than it is
to forget enabling them.