chore: Upgrade sqlxmq to 0.2
Upgrade the sqlxmq to the newly released version 0.2.1. This enables
removing the reqwest client static, and allows for a cleaner way to take
the job listener from the client. The response sender static was not
removed for now as it would have required removing the spawn methods
from requests and only having them in clients.